backend/packages/ai_tax_search/pyproject.toml
[tool.poetry]
name = "ai_tax_search"
version = "0.1.0"
description = "AI Tax project search"
authors = ["Lukasz Augustyniak <aisolutions@lukaszaugustyniak.com>"]
readme = "README.md"
packages = [
    { include = "ai_tax_search" }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
langchain = "^0.3"
langchain-community = "^0.3.19"
langchain-openai = "^0.3.7"
weaviate-client = "^4.11.1"
termcolor = "^2.5.0"
loguru = "^0.7.3"
jinja2 = "^3.1.6"
python-dotenv = "^1.0.0"
langfuse = "^3.2.3"
rich = "^14.1.0"
celery = {extras = ["redis"], version = "^5.5.3"}
jsonschema = "^4.23.0"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
ruff = "^0.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.black]
line-length = 120
target-version = ['py312']

[tool.langserve]
export_module = "ai_tax_search"
export_attr = "chain"
